+++ src/js/edu-benefits/5490/config/form.js
@@ -97,15 +97,7 @@ const formConfig = {
       title: 'Applicant Information',
       pages: {
         applicantInformation: applicantInformation(fullSchema5490, {
-          labels: { relationship: relationshipLabels },
-          fields: [
-            'relativeFullName',
-            'relativeSocialSecurityNumber',
-            'view:noSSN',
-            'relativeDateOfBirth',
-            'gender',
-            'relationship'
-          ]
+          labels: { relationship: relationshipLabels }
         }),
         additionalBenefits: additionalBenefits(fullSchema5490, {
           fields: ['civilianBenefitsAssistance', 'civilianBenefitsSource']
